+++ suppaftp/src/async_ftp/tokio_ftp/tls.rs
@@ -4,9 +4,9 @@
 
 use std::fmt::Debug;
 
+use async_trait::async_trait;
 use tokio::io::{AsyncRead, AsyncWrite, ReadBuf};
 use tokio::net::TcpStream;
-use async_trait::async_trait;
 
 use crate::FtpResult;